ia-nuevo/requirements.txt
quivr-core>=0.0.33
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
xlrd>=2.0.0
pandas>=2.0.0
cffi
pypdf>=3.0.0
numpy<2.0
unstructured[pdf]>=0.10.0
